+++ src/class_diagram/generators/json/class_diagram_generator.cc
@@ -126,6 +126,8 @@ void generator::generate(std::ostream &ostr) const
 
     generate_relationships(json_);
 
+    generate_metadata(json_);
+
     ostr << json_;
 }