+++ src/class_diagram/generators/plantuml/class_diagram_generator.cc
@@ -765,6 +765,8 @@ void generator::generate(std::ostream &ostr) const
 
     generate_plantuml_directives(ostr, m_config.puml().after);
 
+    generate_metadata(ostr);
+
     ostr << "@enduml" << '\n';
 }