+++ src/include_diagram/generators/json/include_diagram_generator.cc
@@ -100,6 +100,8 @@ void generator::generate(std::ostream &ostr) const
         generate_relationships(dynamic_cast<source_file &>(*f), json_);
     });
 
+    generate_metadata(json_);
+
     ostr << json_;
 }
 } // namespace clanguml::include_diagram::generators::json