+++ src/package_diagram/generators/plantuml/package_diagram_generator.cc
@@ -147,6 +147,8 @@ void generator::generate(std::ostream &ostr) const
 
     generate_plantuml_directives(ostr, m_config.puml().after);
 
+    generate_metadata(ostr);
+
     ostr << "@enduml" << '\n';
 }